refactor: Generic な Redux hooks を導入
